HUGE update
I have created a webpage that allows users to input information.
I have also made it so that the team number is saved in the settings of the watch and if the user hasn't selected a team, it is defaulted to 548.
I am currently running into an issue where it says that I am forbidden to access the api so...